Relax required defines
Made the following preprocessor defines optional:

DAVAHI_DAEMON_RUNTIME_DIR
Used as the parent of the the PID file.  If left undefined, we just don't make
or clean up a PID file.  This will still be required if you're building with
chroot support.

AVAHI_SOCKET
This was used for a simpler-than-dbus IPC mechanism over a domain socket.  Now
if left undefined, this simple_protocol code is not compiled in.

AVAHI_CONFIG_DIR
Was used to set up inotify watches to restart on config changes.  Our main
config file will be on a read-only partition.  The services config file is
still being watched via the AVAHI_SERVICE_DIR macro.

AVAHI_DBUS_INTROSPECTION_DIR
Location of the dbus xml files describing the avahi api.  Now if left undefined
we'll fail to respond to an introspection request over dbus.

Change-Id: I509dfa393d6d6f0a4cdaff9fefa01adbf9d4216b